Improve based upon vswarm invoker
