+++ go.mod
@@ -3,6 +3,7 @@ module github.com/eth-easl/easyloader
 go 1.17
 
 require (
+	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20211206145412-3a4607297428
 	github.com/gocarina/gocsv v0.0.0-20211203214250-4735fba0c1d9
 	github.com/golang/protobuf v1.5.2
 	github.com/sirupsen/logrus v1.8.1
@@ -11,6 +12,15 @@ require (
 )
 
 require (
+	github.com/containerd/containerd v1.5.2 // indirect
+	github.com/openzipkin/zipkin-go v0.2.5 // indirect
+	go.opentelemetry.io/contrib v0.20.0 // indirect
+	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.20.0 // indirect
+	go.opentelemetry.io/otel v0.20.0 // indirect
+	go.opentelemetry.io/otel/exporters/trace/zipkin v0.20.0 // indirect
+	go.opentelemetry.io/otel/metric v0.20.0 // indirect
+	go.opentelemetry.io/otel/sdk v0.20.0 // indirect
+	go.opentelemetry.io/otel/trace v0.20.0 // indirect
 	golang.org/x/net v0.0.0-20210405180319-a5a99cb37ef4 // indirect
 	golang.org/x/text v0.3.6 // indirect
 	google.golang.org/genproto v0.0.0-20211208223120-3a66f561d7aa // indirect
